- Moved flashpreview in a separate package - added some doctests - antialias for thumbnails - Detect the filetype at upload and give a specific interface IProjectVideo or IProjectImage

+++ src/eztranet/project/project.py
@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-from zope.interface import implements
+from zope.interface import implements, implementsOnly
 from zope.component import adapts, adapter
 from zope.app.container.interfaces import INameChooser, IObjectRemovedEvent, IObjectAddedEvent, IObjectRemovedEvent
 from zope.lifecycleevent.interfaces import IObjectModifiedEvent
@@ -7,96 +6,88 @@ from zope.app.container.contained import NameChooser
 from zope.app.container.btree import BTreeContainer
 from zope.component.factory import Factory
 from zope.file.file import File
-from zope.app.file.image import Image
 from persistent import Persistent
 import os
 import PIL.Image
 from StringIO import StringIO
 from tempfile import NamedTemporaryFile
 from interfaces import *
-from flashpreview import compute_flashvideo
 from eztranet.thumbnail.interfaces import IThumbnail
+from eztranet.flashpreview.interfaces import IFlashPreview
 
 class ProjectContainer(BTreeContainer):
-  "a project container"
+  """
+  a project container
+  """
   implements(IProjectContainer)
-  __name__=__parent__=None
+  __name__ = __parent__ = None
 
 ProjectContainerFactory = Factory(ProjectContainer)
     
 class Project(BTreeContainer):
     implements(IProject)
-    title=u""
-    description=u""
-    __name__=__parent__=None
-    def __init__(self, title=None, description=None):
-        self.title, self.description = title, description
-        super(Project, self).__init__()
+    __name__ = __parent__ = None
+
+    def __init__(self, title=u'', description=u''):
+        super(Project, self).__init__( )
+        self.title = title
+        self.description = description
 
 ProjectFactory = Factory(Project)
 
-class ProjectItem(Persistent):
+class ProjectItem(File):
+    """
+    A project item is just a blob file
+    """
     implements(IProjectItem)
-    title=u""
-    description=u""
     __name__=__parent__=None
-    def __init__(self, title=None, description=None):
+
+    def __init__(self, title=u'', description=u''):
+        super(ProjectItem, self).__init__()
         self.title = title
         self.description = description
-        super(ProjectItem, self).__init__()
 
 ProjectItemFactory = Factory(ProjectItem)
 
-class ProjectNameChooser(NameChooser):
-    u"""
-    adapter that allows to choose the __name__ of a project
+class ProjectItemNameChooser(NameChooser):
+    """
+    adapter that allows to choose the __name__ of a projectitem
     """
     adapts(IProjectContainer)
     implements(INameChooser)
-    def chooseName(self, name, project):
+
+    def chooseName(self, name, item):
+        rawname = item.title
+        newname = u''
         if name:
-            return name
-        if project is not None and len(project.title)>0:
-            rawname = project.title
-            newname = unicode.lower(rawname).strip().replace(' ','-').replace(u'/',u'-').lstrip('+@')
-            return newname
-        raise "ProjectNameChooser Error"
+            newname = name
+        if item is not None and len(rawname)>0:
+            newname = unicode.lower(rawname).strip(' @+').replace(' ','-').replace('/','-')
+        return super(ProjectItemNameChooser, self).chooseName(newname, item).encode('utf-8')
 
-
-class ProjectItemNameChooser(NameChooser):
-    u"""
-    adapter that allows to choose the __name__ of a projectitem
+class ProjectNameChooser(ProjectItemNameChooser):
+    """
+    adapter that allows to choose the __name__ of a project
     """
     adapts(IProject)
     implements(INameChooser)
-    def chooseName(self, name, projectitem):   
-        if name:
-            return name
-        if projectitem is not None and len(projectitem.title)>0:
-            rawname = projectitem.title
-            newname = unicode.lower(rawname).strip().replace(' ','-').replace(u'/',u'-').lstrip('+@')
-            return newname
-        raise "ProjectItemNameChooser Error"
 
-class ProjectImage(ProjectItem, Image):
-    implements(IProjectImage)
-    __parent__ = __name__ = None
-    title = description = u""
-
-class ProjectVideo(ProjectItem, File):
+class ProjectVideo(ProjectItem):
     implements(IProjectVideo)
-    __parent__ = __name__ = None
-    title = description = u""
-    flash_video = flash_video_tempfile = None
-        
+
+class ProjectImage(ProjectItem):
+    implements(IProjectImage)
+
 class SearchableTextOfProject(object):
-    u"""
-    l'adapter qui permet d'indexer les projects
+    """
+    adapter that allows to index projects
     """
     adapts(IProject)
     implements(ISearchableTextOfProject)
+
     def __init__(self, context):
         self.context = context
+
     def getSearchableText(self):
         sourcetext = texttoindex = (self.context.title or "") + " " + (self.context.description or "")
         for word in sourcetext.split():        
@@ -104,71 +95,71 @@ class SearchableTextOfProject(object):
                 texttoindex += subword + " "
         return texttoindex
 
-class SearchableTextOfProjectItem(object):
-    u"""
-    l'adapter qui permet d'indexer les projects
+class SearchableTextOfProjectItem(SearchableTextOfProject):
+    """
+    adapter that allows to index project items
     """
     adapts(IProjectItem)
-    implements(ISearchableTextOfProjectItem)
-    def __init__(self, context):
-        self.context = context
-    def getSearchableText(self):
-        sourcetext = texttoindex = (self.context.title or "") + " " + (self.context.description or "")
-        for word in sourcetext.split():        
-            for subword in [ word[i:] for i in xrange(len(word)) if len(word)>=1 ]:
-                texttoindex += subword + " "
-        return texttoindex
+    implementsOnly(ISearchableTextOfProjectItem)
 
 
 class ProjectThumbnail(object):
-    u"adapter from a project to a thumbnail"
+    "adapter from a project to a thumbnail"
     adapts(IProject)
     implements(IThumbnail)
     image = None
     url = '/@@/folder.png'
+
     def __init__(self, context):
         self.context = context
+
     def compute_thumbnail(self):
         pass
 
 @adapter(IProjectImage)
-def ProjectImageThumbnailer(image):
-    u"thumbnail creator for ProjectImage"
+def ProjectImageThumbnailer(imagefile):
+    "thumbnail creator for ProjectImage"
     tmp=StringIO()
-    i = PIL.Image.open(StringIO(image.data))
-    i.thumbnail((120,120))
+    i = PIL.Image.open(imagefile.open())
+    i.thumbnail((120,120), PIL.Image.ANTIALIAS)
     i.save(tmp, "png")
     return tmp.getvalue()
 
 @adapter(IProjectVideo)
-def ProjectVideoThumbnailer(video):
-    u"thumbnail creator for ProjectVideo"
-    u"convert the video to png, without audio, with only 1 frame, with a delay of 3 seconds"
+def ProjectVideoThumbnailer(videofile):
+    "thumbnail creator for ProjectVideo"
+    "convert the video to png, without audio, with only 1 frame, with a delay of 3 seconds"
     tmpfile = NamedTemporaryFile()
-    tmpfile.write(video.open().read())
+    tmpfile.write(videofile.open().read())
     tmpfile.flush()
     thumbnail = os.popen("ffmpeg -i %s -vcodec png -ss 3 -vframes 1 -an -f rawvideo -" % tmpfile.name).read()
     tmpfile.close()
-    return ProjectImageThumbnailer(Image(thumbnail))
+    blobfile = File()
+    fd = blobfile.open('w')
+    fd.write(thumbnail)
+    fd.close()
+    return ProjectImageThumbnailer(blobfile)
 
 @adapter(IProjectVideo, IObjectAddedEvent)
 def ProjectVideoAdded(video, event):
-    u"warning, here the object is NOT security proxied"
-    video.flash_video_tempfile = compute_flashvideo(video)
+    "warning, here the object is NOT security proxied"
+    #IThumbnail(video).compute_thumbnail()
+    IFlashPreview(video).encode()
 
 @adapter(IProjectVideo, IObjectModifiedEvent)
 def ProjectVideoModified(video, event):
-    u"warning, here the object IS security proxied"
+    "warning, here the object IS security proxied"
     try:
         if 'data' in event.descriptions[0].attributes:
             # we compute the flash video only if we uploaded something
-            video.flash_video_tempfile = compute_flashvideo(video)
+            IFlashPreview(video).encode()
     except:
         return
 
 @adapter(IProjectVideo, IObjectRemovedEvent)
 def ProjectVideoRemoved(video, event):
-    tmpfile = video.flash_video_tempfile
-    for file in tmpfile, tmpfile+".OK", tmpfile+'.FAILED': 
-        if os.path.exists(file):
-            os.remove(file)
+    tmpfile = IFlashPreview(video).flash_movie
+    if type(tmpfile) is str and tmpfile[0:4] == '/tmp':
+        for file in tmpfile, tmpfile+".OK", tmpfile+'.FAILED': 
+            if os.path.exists(file):
+                os.remove(file)
